A random desktop background changer
This is super bare bones. I wrote it up in a few minutes because it
seemed like a fun idea. There are a few Hieronymus Bosch pieces and
there are a few from Pieter Bruegel the Elder. The program just chooses
a small frame from a random one of the images and sets it as your
desktop background. Right now all I know is that it works on my Windows
10 laptop.